integridad tablas local, tipor y horarioslocales

+++ app/src/main/java/sv/edu/ues/fia/eisi/pdm115/g10/polideportivoues/ControlBDG10William.java
@@ -34,23 +34,20 @@ public class ControlBDG10William {
         DBhelper.close();
     }
 
-    //Metodos para tabla local
+    /*---------------------Metodos para tabla Local--------------*/
     public String ingresarLocal(Local local){
         String localInsertado = "Local ";
         long cuenta = 0;
-
         ContentValues locales = new ContentValues();
         locales.put("idLocal", local.getIdLocal());
         locales.put("nomLocal", local.getNomLocal());
         locales.put("cantidadPersonas", local.getCantidadPersonas());
         cuenta = db.insert("local",null,locales);
-
         if(cuenta == -1 || cuenta == 0){
             localInsertado = "Error al ingresar un local con id que ya existe, Verificar su inserción";
         }else{
             localInsertado = localInsertado + cuenta + " Registrado";
         }
-
         return localInsertado;
     }
 
@@ -81,23 +78,51 @@ public class ControlBDG10William {
         }
     }
 
-    public String eliminarLocal(Local local){
-        String regAfectados="filas afectadas= ";
-        int contador=0;
-        if (verificarIntegridadDeDatos(local,1)) {
-            if (verificarIntegridadDeDatos(local,2)) {
-            return "El local no puede ser eliminado porque existen registros de horarios con este local.";
-            }else{
-            contador+=db.delete("local", "idLocal='"+local.getIdLocal()+"'", null);
-            regAfectados+=contador;
-            return regAfectados;
-            }
-        }else{
-            return "El local no existe";
+
+    public String eliminarLocal(Local local, int tipo){
+        String regAfectados;
+        int contador;
+        switch (tipo){
+            case 0:
+                if (verificarIntegridadDeDatos(local,1)) {
+                    if (verificarIntegridadDeDatos(local,3)) {
+                        return "3";
+                    }else{
+                        if (verificarIntegridadDeDatos(local,2)) {
+                            return "2";
+                        }else {
+                            regAfectados="filas afectadas= ";
+                            contador=0;
+                            contador += db.delete("local", "idLocal='" + local.getIdLocal() + "'", null);
+                            regAfectados += contador;
+                            return regAfectados;
+                        }
+                    }
+                }else{
+                    return "El local no existe";
+                }
+            case 1:
+                regAfectados="filas afectadas= ";
+                contador=0;
+                contador+=db.delete("horariosLocales", "idLocal='"+local.getIdLocal()+"'", null);
+                contador+=db.delete("local", "idLocal='"+local.getIdLocal()+"'", null);
+                regAfectados+=contador;
+                return regAfectados;
+            case 2:
+                regAfectados="filas afectadas= ";
+                contador=0;
+                contador+=db.delete("reservacion", "idLocal='"+local.getIdLocal()+"'", null);
+                contador+=db.delete("horariosLocales", "idLocal='"+local.getIdLocal()+"'", null);
+                contador+=db.delete("local", "idLocal='"+local.getIdLocal()+"'", null);
+                regAfectados+=contador;
+                return regAfectados;
+            default:
+                return null;
         }
+
     }
 
-    //Metodos para tabla Reservacion
+    /*---------------------Metodos para tabla TipoReservacion--------------*/
     public String ingresarTipoReservacion(TipoReservacion tipoReservacion){
         String TipoReservacionInsertado = "Tipo de reservacion ";
         long cuenta = 0;
@@ -117,15 +142,15 @@ public class ControlBDG10William {
     }
 
     public String actualizarTipoReservacion(TipoReservacion tipoReservacion){
-        //if(verificarIntegridadDeDatos(tipoReservacion,2)){
+        if(verificarIntegridadDeDatos(tipoReservacion,4)){
         String[] id={tipoReservacion.getIdTipoR()};
         ContentValues contentValues = new ContentValues();
         contentValues.put("nomTipoR", tipoReservacion.getNomTipoR());
         db.update("tipoReservacion",contentValues, "idTipoR = ?",id);
         return "Tipo de reservacion Actualizado";
-        //}else{
-        //return "El Tipo de reservacion no existe";
-        //}
+        }else{
+        return "El Tipo de reservacion no existe";
+        }
     }
 
     public TipoReservacion consultarTipoReservacion(String idTipoR){
@@ -141,22 +166,37 @@ public class ControlBDG10William {
         }
     }
 
-    public String eliminarTipoReservacion(TipoReservacion tipoReservacion){
-        String regAfectados="filas afectadas= ";
-        int contador=0;
-        //if (verificarIntegridadDeDatos(tipoReservacion,2)) {
-        //if (verificarIntegridadDeDatos(tipoReservacion,3)) {
-        //return "El tipo de reservacion no puede ser eliminado porque existen registros de reservacion con este tipo.";
-        //}else{
-        contador+=db.delete("tipoReservacion", "idTipoR='"+tipoReservacion.getIdTipoR()+"'", null);
-        regAfectados+=contador;
-        return regAfectados;
-        //}
-        //}else{
-        //return "El tipo de reservacion no existe";
-        //}
+    public String eliminarTipoReservacion(TipoReservacion tipoReservacion, int tipo){
+        String regAfectados;
+        int contador;
+        switch (tipo) {
+            case 0:
+                regAfectados="filas afectadas= ";
+                contador=0;
+                if (verificarIntegridadDeDatos(tipoReservacion, 4)) {
+                    if (verificarIntegridadDeDatos(tipoReservacion, 5)) {
+                        return "1";
+                    } else {
+                        contador += db.delete("tipoReservacion", "idTipoR='" + tipoReservacion.getIdTipoR() + "'", null);
+                        regAfectados += contador;
+                        return regAfectados;
+                    }
+                } else {
+                    return "El tipo de reservacion no existe";
+                }
+            case 1:
+                regAfectados="filas afectadas= ";
+                contador=0;
+                contador += db.delete("reservacion", "idTipoR='" + tipoReservacion.getIdTipoR() + "'", null);
+                contador += db.delete("tipoReservacion", "idTipoR='" + tipoReservacion.getIdTipoR() + "'", null);
+                regAfectados += contador;
+                return regAfectados;
+            default:
+                return null;
+        }
     }
 
+    /*---------------------Metodos para tabla HorariosLocales--------------*/
     public String ingresarHorariosLocales(HorariosLocales horariosLocales){
         String localInsertado = "Horario del Local ";
         long cuenta = 0;
@@ -176,7 +216,7 @@ public class ControlBDG10William {
     }
 
     public String actualizarHorariosLocales(HorariosLocales horariosLocales){
-        //if(verificarIntegridadDeDatos(local,1)){
+        if(verificarIntegridadDeDatos(horariosLocales,6)){
             long cuenta;
             String[] id={horariosLocales.getIdHorario(), horariosLocales.getIdLocal()};
             ContentValues contentValues = new ContentValues();
@@ -187,9 +227,9 @@ public class ControlBDG10William {
             }else{
                 return "Horario del local Actualizado";
             }
-        //}else{
-            //return "El local no existe";
-        //}
+        }else{
+            return "El horario del local no existe";
+        }
     }
 
     public HorariosLocales consultarHorariosLocales(String idHorario, String idlocal){
@@ -206,22 +246,37 @@ public class ControlBDG10William {
         }
     }
 
-    public String eliminarHorariosLocales(HorariosLocales horariosLocales){
-        String regAfectados="filas afectadas= ";
-        int contador=0;
-         //if (verificarIntegridadDeDatos(local,1)) {
-            //if (verificarIntegridadDeDatos(local,2)) {
-            //return "El local no puede ser eliminado porque existen registros de local evento con este local.";
-            //}else{
-            contador+=db.delete("horariosLocales", "idHorario = '"+ horariosLocales.getIdHorario()+"' and idLocal = '"+ horariosLocales.getIdLocal()+"'", null);
-            regAfectados+=contador;
-            return regAfectados;
-            //}
-        //}else{
-            //return "El local no existe";
-        //}
+    public String eliminarHorariosLocales(HorariosLocales horariosLocales, int tipo){
+        String regAfectados;
+        int contador;
+        switch (tipo) {
+            case 0:
+                regAfectados = "filas afectadas= ";
+                contador = 0;
+                if (verificarIntegridadDeDatos(horariosLocales, 6)) {
+                    if (verificarIntegridadDeDatos(horariosLocales, 7)) {
+                        return "1";
+                    } else {
+                        contador += db.delete("horariosLocales", "idHorario = '" + horariosLocales.getIdHorario() + "' and idLocal = '" + horariosLocales.getIdLocal() + "'", null);
+                        regAfectados += contador;
+                        return regAfectados;
+                    }
+                } else {
+                    return "El local no existe";
+                }
+            case 1:
+                regAfectados = "filas afectadas= ";
+                contador = 0;
+                contador += db.delete("reservacion", "idHorario='" + horariosLocales.getIdHorario() + "' and idLocal='" + horariosLocales.getIdLocal() + "'", null);
+                contador += db.delete("horariosLocales", "idHorario = '" + horariosLocales.getIdHorario() + "' and idLocal = '" + horariosLocales.getIdLocal() + "'", null);
+                regAfectados += contador;
+                return regAfectados;
+            default:
+                return null;
+        }
     }
 
+    /*---------------------Verificar integridad--------------*/
     private boolean verificarIntegridadDeDatos(Object valor, int relacion) throws SQLException{
         switch (relacion){
             case 1: {
@@ -237,19 +292,79 @@ public class ControlBDG10William {
                 }
             }
             case 2:{
+                //Verifica si existen horarios asociados al local
                 Local local = (Local) valor;
                 Cursor c=db.query(true, "horariosLocales", new String[] {"idLocal"}, "idLocal='"+local.getIdLocal()+"'",null, null, null, null, null);
-                if(c.moveToFirst())
+                if(c.moveToFirst()){
                     return true;
-                else
+                }
+                else{
                     return false;
+                }
+            }
+            case 3: {
+                //verifica si existen reservaciones asociadas a este local
+                Local local = (Local) valor;
+                Cursor c=db.query(true, "reservacion", new String[] {"idLocal"}, "idLocal='"+local.getIdLocal()+"'",null, null, null, null, null);
+                if(c.moveToFirst()){
+                    return true;
+                }
+                else{
+                    return false;
+                }
+            }
+            case 4: {
+                //Verifica si existe el tipo de reservacion a actualizar
+                TipoReservacion tipoR = (TipoReservacion) valor;
+                String[]id = {tipoR.getIdTipoR()};
+                open();
+                Cursor cursor = db.query("tipoReservacion",null,"idTipoR = ?", id,null,null,null);
+                if (cursor.moveToFirst()){
+                    return true;
+                }else{
+                    return false;
+                }
+            }
+            case 5:{
+                //Verifica si existen reservaciones asociadas al tipo
+                TipoReservacion tipoR = (TipoReservacion) valor;
+                Cursor c=db.query(true, "reservacion", new String[] {"idTipoR"}, "idTipoR='"+tipoR.getIdTipoR()+"'",null, null, null, null, null);
+                if(c.moveToFirst()){
+                    return true;
+                }
+                else{
+                    return false;
+                }
+            }
+            case 6:{
+                //Verifica si existe el horario
+                HorariosLocales horariosLocales = (HorariosLocales) valor;
+                String[]id = {horariosLocales.getIdHorario(),horariosLocales.getIdLocal()};
+                Cursor cursor = db.query("horariosLocales",null,"idHorario = ? and idLocal = ?", id,null,null,null);
+                if(cursor.moveToFirst()){
+                    return true;
+                }
+                else{
+                    return false;
+                }
+            }
+            case 7:{
+                //Verifica si existen reservaciones asociadas al horario
+                HorariosLocales horariosLocales = (HorariosLocales) valor;
+                Cursor c=db.query(true, "reservacion", new String[] {"idHorario","idLocal"}, "idHorario='"+horariosLocales.getIdHorario()+"' and idLocal='"+horariosLocales.getIdLocal()+"'",null, null, null, null, null);
+                if(c.moveToFirst()){
+                    return true;
+                }
+                else{
+                    return false;
+                }
             }
             default:
                 return false;
         }
     }
 
-    //Extraemos todas las horas registradas en la base de datos
+    /*---------------------Consultas para los select--------------*/
     public List<HorariosDisponibles> consultarHorarioDisponibles(){
         List<HorariosDisponibles> arrayHorarios=new ArrayList<>();
         Cursor cursor = db.query("horariosDisponibles",null, null, null, null, null, null);
